Minor cleanups.
This PR updates a few parts to be a little more Go-idiomatic
without changing any significant semantics.

Package comments are attached to the package statement so that they'll
show up in godoc.

GoSend is changed to a simple function, which arguably looks slightly
nicer and uses one less allocation per send.

A couple of race conditions cleaned up in graphml.

A few other little tweaks.